+++ consent/models.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Models for edX Enterprise's Consent application.
+"""
+
+from __future__ import unicode_literals