+++ enterprise/urls.py
@@ -39,4 +39,9 @@ urlpatterns += [
         include('integrated_channels.integrated_channel.urls'),
         name='integrated_channel'
     ),
+    url(
+        r'',
+        include('consent.urls'),
+        name='consent'
+    ),
 ]